ai-twin-retina-eyes_V01_2/ai-twin-retina-eyes/requirements.txt
numpy
opencv-python
pyserial
torch
